ex 4.21 bloglist - delete authentication

+++ part4/bloglist-backend/controllers/blogs.js
@@ -33,6 +33,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findOne({ _id: request.params.id })
+    console.log(user, blog)
+    if (user._id.toString() !== blog.user.toString()){
+        return response.status(403).json({ error: 'No permission to operate' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
